requirements.txt
# Core Web Framework
streamlit==1.31.0

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.0
python-docx==1.1.0
docling==1.5.1

# Translation Services
googletrans==4.0.0rc1
deep-translator==1.11.4

# AI/ML and OpenAI
openai==1.12.0
langchain==0.1.8
langchain-openai==0.0.6
langchain-community==0.0.20

# Vector Databases and Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1
faiss-cpu==1.8.0

# RAG and Text Processing
tiktoken==0.5.2
beautifulsoup4==4.12.3
nltk==3.8.1
langdetect==1.0.9

# API and HTTP
requests==2.31.0
aiohttp==3.9.3

# Data Processing
pandas==2.2.0
numpy==1.26.4

# Configuration and Environment
python-dotenv==1.0.1
pydantic==2.6.1

# Neo4j for Persistent Memory (optional)
neo4j==5.17.0

# Additional Utilities
python-magic==0.4.27
chardet==5.2.0
pathlib2==2.3.7

# Development and Testing
pytest==8.0.1
black==24.2.0
flake8==7.0.0 